Some initial work on threads.
The test threads always wait for a specific amount of time.
The time was meant to be a fail safe to ensure the tests actually quit but it is being used as a mechanism for actual tests.
This is the first stage in removing this by adding a continue to the test thread allowing the waithandle to be set and continue processing.

[Ticket: #189]