Implementing chain link id
Adding dependencies for chain links and adding Ids. As a side note, Id for a chain link will ALWAYS be passed and adding it as a header to the inflator feels wrong.
It should be passed in the Create(x) method because every components chain link will receive an Id and it is up to the individual component to decide if it needs / wants to use it.
Passing it as a header means that every chain link could receive it and if there is any validation on those headers then this could potentially cause unexpected errors down the line.

[Ticket: #176]