Rest to and converter (basic) done.
The HttpContentFactory.cs is a basic implementation and it is expected that if the user needs more complex cases, they should overwrite the implementation of this factory in their startup code.

Also moved the stack to use podman because this works better on my machine. Need to use env file to load in the DOCKER_HOST for localstack. This should mean that localstack will run correctly on anyones machine.

[Ticket: #176]